+++ src/mpid/ch4/src/ch4r_proc.c
@@ -212,7 +212,7 @@ int MPIDIU_avt_destroy(void)
 /* convert upid to gpid by netmod.
  * For ofi netmod, it inserts the address and fills an av entry.
  */
-int MPIDIU_upids_to_gpids(int size, size_t * remote_upid_size, char *remote_upids,
+int MPIDIU_upids_to_gpids(int size, int *remote_upid_size, char *remote_upids,
                           uint64_t * remote_gpids)
 {
     int mpi_errno = MPI_SUCCESS;
